+++ server/api/views/assistant/eval_assistant.py
@@ -1,0 +1,15 @@
+# uv script (or plain Python) to generate results to CSV, run from the terminal 
+
+import asyncio
+
+# Set of representative questions
+
+
+# Read model and INSTRUCTIONS from the source file or add a lightweight config endpoint to the backend
+
+
+async def main():
+
+
+if __name__ == "__main__":
+    asyncio.run(main())
